Updated types to match UnifiedDataLoader, PR #1073
